extracted internal item registry
Extract DSP item registry, so it can be maintained separately
