--- src/assets/index.mts
@@ -1,6 +1,0 @@
-export * from "./facility.mts";
-export * from "./recipes/index.mts";
-export * from "./sorter.mts";
-export * from "./proliferator.mts";
-
-export const GAME_VERSION = "0.10.29.21950";
