--- src/assets/recipes/index.mts
@@ -1,12 +1,0 @@
-export {
-	RecipeType,
-	RECIPE_REGISTRY,
-	RECIPE_DEFAULT_LOOKUP,
-} from "./recipe.mts";
-
-import "./assembler.mts";
-import "./chemical.mts";
-import "./particle.mts";
-import "./refining.mts";
-import "./research.mts";
-import "./smelting.mts";
